scripts/requirements.txt
azure-ai-projects>=1.0.0
azure-identity>=1.15.0
